feat: store selected locale in local storage and use value on next visit
